requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_r-mutex=1.0.1=anacondar_1
absl-py=1.4.0=pypi_0
aldy=4.4=pypi_0
asgiref=3.7.2=pypi_0
binutils_impl_linux-64=2.40=hf600244_0
biothings-client=0.3.0=pypi_0
boost=1.78.0=py39h7c9e3ff_4
boost-cpp=1.78.0=h6582d0a_3
bwa=0.7.17=he4a0461_11
bwidget=1.9.14=ha770c72_1
bx-python=0.10.0=pypi_0
bzip2=1.0.8=h7f98852_4
c-ares=1.19.1=hd590300_0
ca-certificates=2023.7.22=hbcca054_0
cairo=1.16.0=hbbf8b49_1016
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
crossmap=0.6.6=pypi_0
curl=8.2.1=hca28451_0
cython=3.0.0=pypi_0
dill=0.3.7=pyhd8ed1ab_0
django=4.2.4=pypi_0
django-cors-headers=4.2.0=pypi_0
djangorestframework=3.14.0=pypi_0
dragmap=1.3.0=h91baf5a_3
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.1.2=pypi_0
expat=2.5.0=hcb278e6_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.12.1=hca18f0e_1
fribidi=1.0.10=h36c2ea0_0
gcc_impl_linux-64=13.1.0=hc4be1a9_0
gettext=0.21.1=h27087fc_0
gfortran_impl_linux-64=13.1.0=hd511a9b_0
graphite2=1.3.13=h58526e2_1001
gsl=2.7=he838d99_0
gxx_impl_linux-64=13.1.0=hc4be1a9_0
harfbuzz=7.3.0=hdb3a94d_0
htslib=1.17=h81da01d_2
icu=72.1=hcb278e6_0
idna=3.4=pypi_0
iniconfig=2.0.0=pypi_0
kernel-headers_linux-64=2.6.32=he073ed8_16
keyutils=1.6.1=h166bdaf_0
krb5=1.21.1=h659d440_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=17_linux64_openblas
libcblas=3.9.0=17_linux64_openblas
libcurl=8.2.1=hca28451_0
libdeflate=1.18=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-devel_linux-64=13.1.0=he3cc6c4_0
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libglib=2.76.4=hebfc3b9_0
libgomp=13.1.0=he5830b7_0
libiconv=1.17=h166bdaf_0
libjpeg-turbo=2.1.5.1=h0b41bf4_0
liblapack=3.9.0=17_linux64_openblas
libnghttp2=1.52.0=h61bc06f_0
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.23=pthreads_h80387f5_0
libpng=1.6.39=h753d276_0
libsanitizer=13.1.0=hfd8a6a1_0
libsqlite=3.42.0=h2797004_0
libssh2=1.11.0=h0841786_0
libstdcxx-devel_linux-64=13.1.0=he3cc6c4_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtiff=4.5.1=h8b53f26_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.1=hd590300_0
libxcb=1.15=h0b41bf4_0
libxml2=2.11.4=h0d562d8_0
libzlib=1.2.13=hd590300_5
logbook=1.6.0=pypi_0
make=4.3=hd18ef5c_1
mappy=2.26=pypi_0
multiprocess=0.70.15=py39hd1e30aa_0
myvariant=1.0.0=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=hcb278e6_0
numpy=1.25.2=pypi_0
openjdk=11.0.1=h470a237_14
openpyxl=3.1.2=pypi_0
openssl=3.1.1=hd590300_1
ortools=9.6.2534=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pango=1.50.14=heaa33ce_1
pcre2=10.40=hc3806b6_0
perl=5.32.1=4_hd590300_perl5
picard=2.18.23=0
pip=23.2.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
pluggy=1.2.0=pypi_0
protobuf=4.23.4=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pybigwig=0.3.22=pypi_0
pysam=0.21.0=pypi_0
pytest=7.4.0=pypi_0
python=3.9.16=h2782a2a_0_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.9=3_cp39
pytz=2023.3=pypi_0
pyvcf=0.6.8=pypi_0
pyyaml=6.0.1=pypi_0
r-base=4.3.1=h29c4799_3
readline=8.2=h8228510_1
requests=2.31.0=pypi_0
samtools=1.17=hd87286a_1
scipy=1.11.1=pypi_0
sed=4.8=he412f7d_0
setuptools=57.5.0=pypi_0
six=1.16.0=pypi_0
sqlparse=0.4.4=pypi_0
sysroot_linux-64=2.12=he073ed8_16
tk=8.6.12=h27826a3_0
tktable=2.10=hb7b940f_3
tomli=2.0.1=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.0.4=pypi_0
wheel=0.41.0=pyhd8ed1ab_0
xlsxwriter=3.1.2=pypi_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.6=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-libxt=1.3.0=hd590300_1
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
zlib=1.2.13=hd590300_5
zstd=1.5.2=hfc55251_7
